+++ code/src/pages/MovieDetails.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const MovieDetails = () => {
+  return;
+  <div>
+    <h1>Movie Detail page</h1>
+  </div>;
+};
+
+export default MovieDetails;
